Add pagination to card news page, showing 15 items at a time
