+++ data/news/2026-02-08-chase-disney-inspire-card.yaml
@@ -5,6 +5,6 @@ summary: "Chase introduces a third Disney-branded credit card, the Disney Inspir
 tags:
   - "new-card"
 bank: "Chase"
-card_slug: "disney-inspire-visa-card"
+card_slug: "disney-inspire-visa"
 source: "Upgraded Points"
 source_url: "https://upgradedpoints.com/news/chase-launches-disney-inspire-visa-card/"
